Ensured DB is UTF-8 case insensitive
Changed the migrations to ensure that the DB columns are stored as utf-8
irrespective of the default server settings, and compared using
case-insensitive collation which is appropriate for all of our needs at
the moment (one specific case being email address lookup). Also added a
test to ensure that email addresses are case-insensitive.