Tests for user info in token responses
As in the updated specification, basic user info should now come back in
the OAuth token responses to aid the client apps in displaying this
information, and help them keep it up to date. Tests now check that this
is returned (note: the tests for the client info are commented out as
this work isn't done yet).