Cleaned up test env and made in proc default
Moved all the test configuration into a TEST_CONFIG constant (like Alex
has done in the api-test library) and defaulted to using in process
testing if the server is on a loopback IP address.

Also changed the starting of the in-process server so that it doesn't
need any checks before each test, but just starts at the start of the
run if it's in process.

Updated the README to reflect the latest situation.